refactor(member): MemberCreateCommand 생성
